refactor: 메소드 이름 변경, 패키지 옮김, 입력 검증 추가
