+++ src/main/java/racingcar/view/RacingGameView.java
@@ -8,9 +8,17 @@ import java.util.Map;
 public class RacingGameView {
     public List<String> inputCarNames() {
         System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
-        return Arrays.stream(Console.readLine().split(","))
+        List<String> inputNames = Arrays.stream(Console.readLine().split(","))
                 .map(String::trim)
                 .toList();
+        validateInputNames(inputNames);
+        return inputNames;
+    }
+
+    private void validateInputNames(List<String> inputNames) {
+        if (inputNames.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
     }
 
     public int inputAttempt() {
