+++ src/test/java/racingcar/model/CarTest.java
@@ -1,11 +1,11 @@
-package racingcar;
+package racingcar.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import org.junit.jupiter.api.Test;
-import racingcar.model.Car;
+import racingcar.util.FixedNumberGenerator;
 import racingcar.util.NumberGenerator;
 import racingcar.util.RandomNumberGenerator;
